Refactor test-utils to be included in tests
