+++ package.json
@@ -1,6 +1,9 @@
 {
   "name": "vaccine-management-frontend",
   "jest": {
+    "transformIgnorePatterns": [
+      "/node_modules/(?!@codemirror)/"
+    ],
     "coveragePathIgnorePatterns": [
       "<rootDir>/src/misc/",
       "<rootDir>/src/types/",
@@ -52,7 +55,7 @@
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
-    "test": "react-scripts test --transformIgnorePatterns \"node_modules/(?!@codemirror)/\"",
+    "test": "react-scripts test",
     "eject": "react-scripts eject"
   },
   "eslintConfig": {
